% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article.R
\name{chiaro_md}
\alias{chiaro_md}
\title{A function for article classes}
\usage{
chiaro_md(fig_width = 2, fig_height = 2, fig_crop = TRUE,
  latex_engine = "xelatex", dev = "tikz", plot_font = "mainfont")
}
\description{
To use scuro's font and figure-drawing setup with slides, use \code{\link{scuro_md}} but with format option \code{scuro=FALSE}.
}

