% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotting.R
\name{dark_plot_theme}
\alias{dark_plot_theme}
\title{A dark ggplot2 theme suitable for slides}
\usage{
dark_plot_theme(base_size = 9, base_family = "", dark = "gray10",
  light = "white")
}
\arguments{
\item{base_size}{general text-size parameter in ?? units}

\item{base_family}{font family; to be overriden in tikz graphics}

\item{dark}{main background color (normally, match slide background)}

\item{light}{main foreground color (normally, match slide text)}
}
\value{
a \code{\link[ggplot2]{theme}} object that can be added to a plot.
  Override settings by adding another \code{theme()} call.
}
\description{
Add this to your ggplots.
}

