% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{make}
\alias{make}
\title{Convert scuro markdown to PDF slides, notes, scripts, and handouts}
\usage{
make(target = "all")
}
\arguments{
\item{target}{Which PDFs to make: for each file \code{notes/X.md} or
\code{scripts/X.md}, the options are \code{slides/X.pdf} (slides),
\code{lectures/X.pdf} (speaker notes), \code{handouts/X.pdf} (slide handout
for the audience). The default, \code{all}, generates all three for each
markdown file.}
}
\description{
Knitted markdown can be converted into the multiple forms that support a
lecture with this function. It is just a wrapper for an invocation of
\code{make}.
}
\details{
The possible targets are defined by the Makefile included in this package
(\emph{not} the Makefile in the \code{lectures} template), which can be
found at \code{system.file("elsmd/Makefile", package="scuro")}.
}

