% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting.R
\name{output_hook_routput}
\alias{output_hook_routput}
\title{Code output environment hook}
\usage{
output_hook_routput(x, options)
}
\description{
Not called directly but automatically set as knitr's output hook. This puts R
output into an \code{ROutput} environment which it is up to you to define
with something like 
\code{\\DefineVerbatimEnvironment{ROutput}{Verbatim}{frame=single,fontsize=\\footnotesize}} (which is what's used in the included \code{talk} RMarkdown template).
}

