% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitting.R
\name{plot_hook_textpos}
\alias{plot_hook_textpos}
\title{Plot placement and resizing hook}
\usage{
plot_hook_textpos(x, options)
}
\description{
Not called directly but automatically set as knitr's plot hook.
Allows you to use the textpos grid for placing and sizing plots on
slides. Example usage below.
}
\details{
This gives us some new options and changes the meaning of two others:

out.width, out.height: reinterpreted as units on the textpos grid.
N.B. this sets the scaling by pdflatex of an image file whose
generated dimensions are set by fig.width and fig.height (always
inches).

textblock_width: width of the block in textpos grid units

textblock_pos: 2-element vector of grid coordinates of the textblock

center: Boolean: enclose graphic within center environment?

inside_textblock: don't generate textblock environment, just the
\code{\\includegraphics} line (useful if, e.g., you write out the environment
\code{\\begin} and \code{\\end} yourself and stick the code chunk between them)

Grid positioning only happens if inside_textblock=T or textblock_width
is specified. Otherwise we revert to knitr's normal behavior. Grid
positioning also always makes graphics inline. Use inside_textblock=T
and enclose it in a figure environment yourself if you want to combine
the textblock and the figure environment (you probably don't).

The main use for this is probably to place a single figure on the grid
instead of in the normal flow of elements.

N.B. the default knitr plot output hook for LaTeX responds to many
more options than this does. But in my workflow we are knitting to
markdown, not LaTeX. That's a simplifying decision in some ways but
with some costs in flexibility.
}

