% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitting.R
\name{print_lm}
\alias{print_lm}
\title{A shortcut for printing out linear models as tables}
\usage{
print_lm(...)
}
\arguments{
\item{...}{passed on to \code{stargazer}. The first parameter is the model.}
}
\value{
LaTeX table code marked with class \code{asis_output}
}
\description{
This routine wraps \code{\link[stargazer]{stargazer}}, adjusting some
options and ensuring as-is output when knitting.
}

