% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitting.R
\name{print_tabular}
\alias{print_tabular}
\title{Print LaTeX commands for typesetting a data frame as a table}
\usage{
print_tabular(x, digits = 0, alignment = paste(ifelse(sapply(x, is.numeric),
  "r", "l"), collapse = ""), include.colnames = T, floating = F,
  caption = NULL, label = NULL, ...)
}
\arguments{
\item{x}{a data frame. Convert matrices and contingency tables yourself.}

\item{digits}{the number of digits after the decimal point.}

\item{alignment}{a single-element character vector giving the sequence of
column alignments according to the \code{tabular} syntax in LaTeX (\code{l, c, r}
for left, center, right, and \code{p{dim}} for text wrapped in a box of
width \code{dim}---for example \code{p{2 in}}). By default numeric columns are
right aligned and the rest are left aligned.}

\item{include.colnames}{whether to print the data frame column names
as column headers. It often makes sense to assign something
human-friendly to the column names first. This function assumes you never
want to print rownames. Keeping anything interesting in the rownames
of a data frame is a bad idea anyway. If you want rownames, add a data
column on the front instead.}

\item{floating}{whether just to drop the table in the flow of text
or to put it in a LaTeX environment like \code{table}. For slides, \code{floating=T} is
probably what you want; beamer centers
floating tables on slides as you'd hope.}

\item{caption}{table caption (provided \code{floating=T})}

\item{label}{the LaTeX label of the table (if \code{floating=T}), so that you can
refer to the table number---but why do that in a talk?}

\item{...}{the rest of the parameters are passed on to \code{\link[xtable]{print.xtable}}.}
}
\description{
This is a helper function for using \code{xtable} in a knitting environment.
It wrangles the bizarre division of options between
\code{\link[xtable]{xtable}} and \code{\link[xtable]{print.xtable}} and
ensures as-is output for knitting.
}

