% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ragoldstBeamer-package.R
\docType{package}
\name{ragoldstBeamer}
\alias{ragoldstBeamer}
\alias{ragoldstBeamer-package}
\title{Miscellaneous R Utilities for Beamer Presentations}
\description{
This package includes some utilities for styling and laying out beamer presentations.
}

