% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{render_pdf}
\alias{render_pdf}
\title{Convert scuro markdown to PDF slides, notes, scripts, and handouts}
\usage{
render_pdf(input, type = c("slides", "handouts", "lectures", "notes"))
}
\arguments{
\item{input}{source markdown (\emph{not} R markdown) file, created by
\code{\link[rmarkdown]{render}}ing to \code{{\link{scuro_md}}}.}

\item{type}{type of PDF to generate (any of \code{slides},
\code{handout[s]}, or \code{lecture[s]} (synonym: \code{notes})). Can be a
vector. Output goes to a correspondingly named subdirectory.}
}
\description{
Knitted markdown can be converted into the multiple forms that support a
lecture with this function. It is just a wrapper for an invocation of
\code{make}.
}
