% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scuro-package.R
\docType{package}
\name{scuro}
\alias{scuro}
\alias{scuro-package}
\title{Miscellaneous R Utilities for Beamer Presentations}
\description{
This package includes some utilities for styling and laying out beamer presentations.
}

