% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{slides}
\alias{handout}
\alias{notes}
\alias{scuro_format}
\alias{slides}
\title{Presentation output formats}
\usage{
slides(...)

notes(...)

handout(...)

scuro_format(output = c("slides", "notes", "handout"), fig_width = 4.5,
  fig_height = 2.75, fig_crop = TRUE, fig_caption = TRUE, dev = "tikz",
  highlight = "zenburn", keep_tex = FALSE, latex_engine = "xelatex",
  pandoc_args = NULL)
}
\arguments{
\item{output}{one of \code{"slides"}, \code{"notes"}, or \code{"handout"}}

\item{fig_width, fig_height}{default figure dimensions (inches). Note that
beamer slides have small "physical" dimensions: 128 mm x 96 mm or about 5
in x 3.75 in. The default values here are meant for slides with just one
figure on them (and possible a title and a caption).}

\item{fig_crop, fig_caption}{same as for the built-in
\code{\link[rmarkdown]{beamer_presentation}}}

\item{dev}{graphics device. \code{"tikz"} for TikZ graphics is the default,
since one of the main points of this package is to transcend bad beamer
fonts. This requires installing the \pkg{tikzDevice} package.}

\item{highlight}{code syntax highlighting. I recommend "zenburn" (see
\code{\link[rmarkdown]{beamer_presentation}} for other choices).}

\item{keep_tex}{whether to keep the intermediate LaTeX file. Often needed for
debugging errors.}

\item{latex_engine}{I recommend xelatex (cf. above remarks about fonts).}

\item{pandoc_args}{character vector of extra arguments to pandoc.}
}
\description{
This package supplies three \pkg{rmarkdown} output formats,
\code{scuro_format::slides}, \code{scuro_format::notes},
and \code{scuro_format::handout}, which can be used in place of
\code{output: beamer_presentation} in the YAML block of an Rmd file or as the
\code{output_format} argument to \pkg{rmarkdown}'s
\code{\link[rmarkdown]{render}} function. The options these formats accept
are given by the parameters to these functions. \code{slides}, \code{notes},
and \code{handout} pass their arguments on to \code{scuro_format}.
}

