% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{scuro_md}
\alias{scuro_md}
\title{Intermediate markdown format}
\usage{
scuro_md(fig_width = 4.5, fig_height = 2.75, dev = "tikz")
}
\arguments{
\item{fig_width, fig_height}{default figure dimensions (inches). Note that
beamer slides have small "physical" dimensions: 128 mm x 96 mm or about 5
in x 3.75 in. The default values here are meant for slides with just one
figure on them (and possibly a title and a caption).}

\item{dev}{graphics device. \code{"tikz"} for TikZ graphics is the default,
since one of the main points of this package is to transcend bad beamer
fonts. This requires installing the \pkg{tikzDevice} package.}

\item{fig_crop, fig_caption}{same as for the built-in
\code{\link[rmarkdown]{beamer_presentation}}}
}
\description{
Intermediate markdown format
}

