% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dark_plot_theme}
\alias{dark_plot_theme}
\title{A dark ggplot2 theme suitable for slides}
\usage{
dark_plot_theme(base_size = 9, base_family = "", dark = "gray10",
  light = "white")
}
\arguments{
\item{base_size}{general text-size parameter in points (??)}

\item{base_family}{font family; to be overriden in TikZ graphics}

\item{dark}{main background color (should probably match slide background).
Some intermediate colors will be interpolated between this and \code{light}.}

\item{light}{main foreground color (should probably match slide text)}
}
\value{
a \code{\link[ggplot2]{theme}} object that can be added to a plot.
  Override settings by adding (or \code{\%replace+}ing) another
\code{theme()} call.
}
\description{
Add this to your ggplots for a light-on-dark theme, suitable for matching to light-on-dark slides. Or use the chunk option \code{dark_theme=T}.
}
