% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/knitting.R
\name{output_hook_routput}
\alias{output_hook_routput}
\title{Code output environment hook}
\usage{
output_hook_routput(x, options)
}
\description{
Not called directly but automatically set as knitr's output hook.
}

