% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting.R
\name{output_hook_routput}
\alias{output_hook_routput}
\title{Code output environment hook}
\usage{
output_hook_routput(x, options)
}
\description{
Not called directly but automatically set as knitr's output hook. This puts
R output into an \code{ROutput} environment. If you use the included
templates this will be set up with a default style. You can adjust the
appearance of this output by adding an array of \code{routput-options} to
your document metadata. These are supplied to the definition of
\code{ROutput} via the \code{fancyvrb} package:
}
\details{
\code{\\DefineVerbatimEnvironment{ROutput}{Verbatim}{...}}

\code{...} is replaced with your \code{routput-options}. If you do not use
the included templates, either remove this output hook or define your own
\code{ROutput} environment.
}

