% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmarkdown.R
\name{scuro_md}
\alias{scuro_md}
\title{Intermediate markdown format}
\usage{
scuro_md(fig_width = 4.5, fig_height = 2.75, fig_crop = TRUE,
  latex_engine = "xelatex", dev = "tikz", plot_font = "sansfont")
}
\arguments{
\item{fig_width, fig_height}{default figure dimensions (inches). Note that
beamer slides have small "physical" dimensions: 128 mm x 96 mm or about 5
in x 3.75 in. The default values here are meant for slides with just one
figure on them (and possibly a title and a caption).}

\item{fig_crop, latex_engine}{same as for the built-in
\code{\link[rmarkdown]{beamer_presentation}}, but XeLaTeX is the default
engine}

\item{dev}{graphics device. \code{"tikz"} for TikZ graphics is the default,
since one of the main points of this package is to transcend bad beamer
fonts. This requires installing the \pkg{tikzDevice} package.}

\item{plot_font}{if TikZ and XeLaTeX are used, the name of the main font for
graphics. If \code{"sansfont"} (the default) or \code{"mainfont"}, the value
of the corresponding document metadata field is used. If unspecified, the
sans font is used if that has been specified. Otherwise we fall back to the
LaTeX default.}
}
\value{
an R Markdown output format used by
  \code{\link[rmarkdown]{render}}.
}
\description{
The strategy adopted by this package is to knit R markdown to markdown
and then control the subsequent PDF generation using a Makefile to
govern pandoc and latex. That lets us generate multiple PDFs from a single
R markdown source without re-knitting.
}
\seealso{
The included \code{lectures} R Markdown template for supporting
files and example source. In RStudio, choose "File > New File > R
Markdown...", choose "From Template," and select "Dark-on-Light Beamer
Slides."
}

