% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knitting.R
\name{tikz_setup_hook}
\alias{tikz_setup_hook}
\title{TikZ and XeLaTeX wrangling hook}
\usage{
tikz_setup_hook(before, options, envir)
}
\description{
This hook ensures that the \pkg{tikzDevice} package options are correctly
set to use xelatex and the chosen font (chunk option \code{plotfont},
normally specified by the document metadata either explicitly or implicitly)
to generate graphics. It is set as the \code{tikz_xelatex} hook and by
default all chunks have \code{tikz_xelatex=T} among their options.
}
